Stack necessary, but optmization still possible
Turns out a stack may still be necessary because the machine wasn't able
to always go back to the right node, as I naively expected. After
thinking a little bit, I couldn't come up with a way to properly guide
the machine to normalizing nodes* other than storing them. But,
regardless, the reclaiming-used-nodes optimization is still possible if,
instead of storing the ports 1 and 2 of normalizing nodes, we store its
port 0. The algorithm is now cleaner and as fast as the optimized version.

To test if this implementation is correct, I've ran an exhaustive search
through the first million λ-terms and compared it to the older stable
version. Turns out this term:

/#a #b /#c b //a a b #a /a a

Was conflitant. But the problem was actually on the older version,
which, for some reason, failed to reduce that term. That means it had a
bug, but adds confidence that this version is correct.

* A normalizing node is a node that has been found out to be part of the
normal form of the graph, and now must explore its ports 1 and 2 to find
more normalizing nodes.